% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehmer.R
\name{lcg.rng}
\alias{lcg.rng}
\title{Create uniform random numbers using Linear Congruential Generator}
\usage{
lcg.rng(x0, a, c, m, iterations = 1)
}
\arguments{
\item{x0}{Initial seed value}

\item{a}{Multiplier}

\item{c}{Shift}

\item{m}{Modulo}

\item{iterations}{Number of iterations}
}
\value{
vector of doubles
}
\description{
Create uniform random numbers using Linear Congruential Generator
}
\examples{
lcg.rng(6, 3, 9, 275)
lcg.rng(6, 3, 9, 275, 20)
}
