% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehmer.R
\name{lcg.seq}
\alias{lcg.seq}
\title{Create sequence of integers from Linear Congruential Generator}
\usage{
lcg.seq(x0, a, c, m, iterations = 1)
}
\arguments{
\item{x0}{Initial seed value}

\item{a}{Multiplier}

\item{c}{Shift}

\item{m}{Modulo}

\item{iterations}{Number of iterations}
}
\value{
vector of integers
}
\description{
Create sequence of integers from Linear Congruential Generator
}
\examples{
lcg.seq(6, 3, 9, 275)
lcg.seq(6, 0, 9, 900, 20)
}
