% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehmer.R
\name{lehmer.rng}
\alias{lehmer.rng}
\title{Create uniform random numbers using Lehmer function}
\usage{
lehmer.rng(x0, a, m, iterations = 1)
}
\arguments{
\item{x0}{Initial seed value}

\item{a}{Multiplier}

\item{m}{Modulo}

\item{iterations}{Number of iterations}
}
\value{
vector of doubles
}
\description{
Create uniform random numbers using Lehmer function
}
\examples{
lehmer.rng(5, 2, 150, 25)
lehmer.rng(5, 2, 150)
}
