% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehmer.R
\name{lehmer.seq}
\alias{lehmer.seq}
\title{Create sequence of integers from Lehmer Random Number Generator}
\usage{
lehmer.seq(x0, a, m, iterations = 1)
}
\arguments{
\item{x0}{Initial seed value}

\item{a}{Multiplier}

\item{m}{Modulo}

\item{iterations}{Number of iterations}
}
\value{
vector of integers
}
\description{
Create sequence of integers from Lehmer Random Number Generator
}
\examples{
lehmer.seq(x0=5, 2, 150)
lehmer.seq(x0=5, 2, 150, iterations=10)
}
